Namespace three.js renderer instead of using prefixes
